add #'PARSE-DECLARATION-IN-BODY which parses possible declarations in a body and returns the rest of the body and the parsed declarations.
add slot FREEP to class SYM that indicates whether a variable is free or bound.
rename slot TYPE in class SYM to DECLSPECS. Its new meaning is the list of DECLSPECs that apply to the symbol.
rename class FORM-WITH-BODY to BODY-FORM to be consistent with naming of the other classes.
add a new class BINDINGS that encapsulates a binding in LET, LET*. rename slot DEFINITION of classes LET-FORM and LET*-FORM to BINDINGS.
add a slot DECLSPECS to classes LET-FORM, LET*-FORM.
change #'PARSE to create BINDINGs instead of a list of SYMs and add parsing declarations in forms LET and LET*.