rename some functions and parameters for easier access.
remove #'DEFAULT-NAMESPACE-LOOKUP/CREATE from exported functons, because it is only a helper function for #'NAMESPACE-LOOKUP/CREATE.
rename #'MAKE-DEFAULT-FREE-COMMON-LISP-NAMESPACE to #'MAKE-FREE-NAMESPACE and have it accept a list of variable, function, and macro to be defined in the free namespace.
rename #'MAKE-CUSTOM-PARSER to #'MAKE-PARSER, and have it accept one argument instead of two.
rename the argument REPARSE and DECLSPEC-REPARSE to PARSER and DECLSPEC-PARSER in #'WALKER:PARSE and #'WALKER-PLUS:PARSE.
rename #'PARSE-WITH-EMPTY-NAMESPACES to #'PARSE-WITH-NAMESPACE, rename the argument REPARSE and DECLSPEC-REPARSE to PARSER and DECLSPEC-PARSER, remove argument FREE-COMMON-LISP-NAMESPACE, and add argument FREE-NAMESPACE.
rename #'NAMESPACES-AT to #'NAMESPACE-AT.
add the constants +COMMON-LISP-MACROS+, +COMMON-LISP-FUNCTIONS+, +COMMON-LISP-VARIABLES+, each defining the list of Common Lisp macros, functions, and variables defined by default.